Pridana security do vsech admin + overeni na deletepost pokud nekdo by chtel nekomu pres ID mazat prispevky

+++ admin/delete.php
@@ -2,17 +2,22 @@
 
 include "../db/db.php";
 
-if(isset($_GET) && is_numeric($_GET["id"]) && is_numeric($_GET["page"])){
-    $id = $_GET['id'];
-    $page = $_GET['page'];
-    $sql = "DELETE FROM cms_users WHERE ID = :id";
-    $ins = [
-        ":id" => $id
-    ];
-    $con = $db->prepare($sql);
-    $con->execute($ins);
-    var_dump($_GET);
-    header("location: index.php?p=".$page."");
+if (isset($_SESSION['admin']) && isset($_SESSION['LOGGED'])) {
+    if ($_SESSION['admin'] == 1 && $_SESSION['LOGGED'] == true) {
+        if(isset($_GET) && is_numeric($_GET["id"]) && is_numeric($_GET["page"])){
+            $id = $_GET['id'];
+            $page = $_GET['page'];
+            $sql = "DELETE FROM cms_users WHERE ID = :id";
+            $ins = [
+                ":id" => $id
+            ];
+            $con = $db->prepare($sql);
+            $con->execute($ins);
+            var_dump($_GET);
+            header("location: index.php?p=".$page."");
+        }
+    }
 }
+
 
 ?>
